requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
pandas
pytz
python-dateutil
reportlab
python-dotenv
